+++ include/uvh5/uvh5_toml.h
@@ -12,7 +12,7 @@ static inline int uvh5_toml_error(const char* msg, const char* msg1)
 		return 1;
 }
 
-int uvh5_toml_sexagesimal_in(toml_table_t* parent, const char* location, float* sexagesimal);
+int uvh5_toml_sexagesimal_in(toml_table_t* parent, const char* location, double* sexagesimal);
 int uvh5_toml_string_in(toml_table_t* parent, const char* location, char** string_out);
 int uvh5_toml_nstring_in(toml_table_t* parent, const char* location, char* string_out, size_t length);
 int uvh5_toml_double_in(toml_table_t* parent, const char* location, double* double_out);
@@ -29,7 +29,7 @@ int uvh5_toml_antenna_table_in(
 	toml_table_t* parent,
 	int* ant_id,
 	char** ant_name,
-	float* ant_pos,
+	double* ant_pos,
 	float* ant_diameter
 );
